timer working now

+++ kernel/globals.c
@@ -7,4 +7,4 @@
 
 int user_setup_stack_ptr;
 int irq_setup_stack_ptr;
-int global_timer;
+unsigned long global_timer;
